# server/main.py
import os
from typing import Optional
import uvicorn
from fastapi import FastAPI, File, Form, HTTPException, Depends, Body, UploadFile
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from fastapi.staticfiles import StaticFiles
from loguru import logger

from models.api import (
    DeleteRequest,
    DeleteResponse,
    GitSearchAddRequest,
    GitSearchAddResponse,
    GitSearchListRequest,
    GitSearchListResponse,
    GitSearchResetDbResponse,
    GitSearchRequest,
    GitSearchResponse,
    QueryRequest,
    QueryResponse,
    SerpSearchARunRequest,
    SerpSearchARunResponse,
    SerpSearchRunRequest,
    SerpSearchRunResponse,
    UpsertRequest,
    UpsertResponse,
)
from datastore.factory import get_datastore
from services.file import get_document_from_file
from server.tools.git_search import GitSearch
from server.tools.web.serp_search import SerpSearch
from models.models import DocumentMetadata, Source
from dotenv import load_dotenv
load_dotenv()

bearer_scheme = HTTPBearer()
BEARER_TOKEN = os.environ.get("BEARER_TOKEN")
assert BEARER_TOKEN is not None


def validate_token(credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme)):
    if credentials.scheme != "Bearer" or credentials.credentials != BEARER_TOKEN:
        raise HTTPException(status_code=401, detail="Invalid or missing token")
    return credentials


app = FastAPI(dependencies=[Depends(validate_token)])
app.mount("/.well-known", StaticFiles(directory=".well-known"), name="static")

# Create a sub-application, in order to access just the query endpoint in an OpenAPI schema, found at http://0.0.0.0:8000/sub/openapi.json when the app is running locally
sub_app = FastAPI(
    title="Retrieval Plugin API",
    description="A retrieval API for querying and filtering documents based on natural language queries and metadata",
    version="1.0.0",
    servers=[{"url": "https://your-app-url.com"}],
    dependencies=[Depends(validate_token)],
)
app.mount("/sub", sub_app)


@app.post(
    "/upsert-file",
    response_model=UpsertResponse,
)
async def upsert_file(
    file: UploadFile = File(...),
    metadata: Optional[str] = Form(None),
):
    try:
        metadata_obj = (
            DocumentMetadata.parse_raw(metadata)
            if metadata
            else DocumentMetadata(source=Source.file)
        )
    except:
        metadata_obj = DocumentMetadata(source=Source.file)

    document = await get_document_from_file(file, metadata_obj)

    try:
        ids = await datastore.upsert([document])
        return UpsertResponse(ids=ids)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail=f"str({e})")


@app.post(
    "/upsert",
    response_model=UpsertResponse,
)
async def upsert(
    request: UpsertRequest = Body(...),
):
    try:
        ids = await datastore.upsert(request.documents)
        return UpsertResponse(ids=ids)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.post(
    "/query",
    response_model=QueryResponse,
)
async def query_main(
    request: QueryRequest = Body(...),
):
    try:
        results = await datastore.query(
            request.queries,
        )
        return QueryResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@sub_app.post(
    "/query",
    response_model=QueryResponse,
    # NOTE: We are describing the shape of the API endpoint input due to a current limitation in parsing arrays of objects from OpenAPI schemas. This will not be necessary in the future.
    description="Accepts search query objects array each with query and optional filter. Break down complex questions into sub-questions. Refine results by criteria, e.g. time / source, don't do this often. Split queries if ResponseTooLargeError occurs.",
)
async def query(
    request: QueryRequest = Body(...),
):
    try:
        results = await datastore.query(
            request.queries,
        )
        return QueryResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.delete(
    "/delete",
    response_model=DeleteResponse,
)
async def delete(
    request: DeleteRequest = Body(...),
):
    if not (request.ids or request.filter or request.delete_all):
        raise HTTPException(
            status_code=400,
            detail="One of ids, filter, or delete_all is required",
        )
    try:
        success = await datastore.delete(
            ids=request.ids,
            filter=request.filter,
            delete_all=request.delete_all,
        )
        return DeleteResponse(success=success)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")

@app.post(
    "/git-search",
    response_model=GitSearchResponse,
)
async def git_search(
    request: GitSearchRequest = Body(...),   
):
    if not request.url:
        raise HTTPException(
            status_code=400,
            detail="url is required",
        )
    try:
        url = request.url
        if hasattr(request, "query"):
            query = request.query
            results = await git_search(url, query)
        else:
            results = await git_search(url)
        return GitSearchResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.post(
    "/list",
    response_model=GitSearchListResponse,
)
async def list(
    request: GitSearchListRequest = Body(...),
):
    try:
        git_search = GitSearch()
        include_files = hasattr(request, "include_files") and request.include_files or False
        results = await git_search.list(include_files)
        print(results)
        return GitSearchListResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.post(
    "/add",
    response_model=GitSearchAddResponse,
)
async def add(
    request: GitSearchAddRequest = Body(...),
):
    print(request)
    
    if not request.urls:
        raise HTTPException(
            status_code=400,
            detail="url is required",
        )
    try:
        urls = request.urls
        filter = hasattr(request, "filter") and request.filter or None
        git_search = GitSearch()
        total_added = await git_search.add(urls, filter)
        print(f"total_added: {total_added}")
        response = GitSearchAddResponse(total_added=total_added)
        print(response)
        return response
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.delete(
    "/reset-db",
    response_model=GitSearchResetDbResponse,
)
async def reset_db():
    try:
        git_search = GitSearch()
        success = await git_search.reset_db()
        return GitSearchResetDbResponse(success)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")

@app.post(
    "/search-run",
    response_model=SerpSearchRunResponse,
    description="Endpoint for running a search query using SerpSearch. Accepts a request object containing the query to run and returns the search results.",

)
def search_run(
    request: SerpSearchRunRequest = Body(...),
):
    """
    Endpoint for running a search query using SerpSearch.

    Args:
        request (SerpSearchRunRequest): The request object containing the query to run.

    Returns:
        SerpSearchRunResponse: The response object containing the search results.
    """
    if not request.query:
        raise HTTPException(
            status_code=400,
            detail="query is required",
        )

    try:
        query = request.query
        serp_search = SerpSearch()
        results = serp_search.run(query=query)
        return SerpSearchRunResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")

@app.post(
    "/search-arun",
    response_model=SerpSearchARunResponse,
)
async def search_arun(
    request: SerpSearchARunRequest = Body(...),
):

    if not request.query:
        raise HTTPException(
            status_code=400,
            detail="query is required",
        )

    try:
        query = request.query
        serp_search = SerpSearch()
        results = await serp_search.arun(query=query)
        return SerpSearchRunResponse(results=results)
    except Exception as e:
        logger.error(e)
        raise HTTPException(status_code=500, detail="Internal Service Error")


@app.on_event("startup")
async def startup():
    global datastore
    datastore = await get_datastore()


def start():
    uvicorn.run("server.main:app", host="0.0.0.0", port=8000, reload=True)